backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic[email]
PyJWT>=2.8.0
tzdata